+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# [1.0.5] - 2020-07-16T17:03+00:00
+
 Changes:
   * Fix problem in _Actions_ dashboard causing 400s using server searches
   * Fix bug causing saved search to default to _All Actions_ on _Actions_ dashboard